% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{additives}
\alias{additives}
\title{Additives}
\format{A dataset with 2499 rows and 7 columns}
\usage{
additives
}
\description{
A dataset containing the union list of approved additives
}
\keyword{datasets}
