% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-functions.R
\name{bracketonly}
\alias{bracketonly}
\title{Cells that are just bracketed numbers - precede with ' to avoid excel turning
it into a negative number}
\usage{
bracketonly(x)
}
\arguments{
\item{x}{The string to be converted}
}
\description{
Cells that are just bracketed numbers - precede with ' to avoid excel turning
it into a negative number
}
