% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-functions.R
\name{counttables}
\alias{counttables}
\title{Get number of tables}
\usage{
counttables(url)
}
\arguments{
\item{url}{URL of the page you want to count the tables on}
}
\description{
Get number of tables
}
