% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-functions.R
\name{decimals}
\alias{decimals}
\title{Euro to British decimals
Converts comma decimal numbers to dot decimal numbers
To be used judiciously on carefully selected columns - it could turn things
like flavouring names into nonsense!}
\usage{
decimals(x)
}
\arguments{
\item{x}{The string to be converted}
}
\description{
Euro to British decimals
Converts comma decimal numbers to dot decimal numbers
To be used judiciously on carefully selected columns - it could turn things
like flavouring names into nonsense!
}
