% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-functions.R
\name{extraspace}
\alias{extraspace}
\title{Clean up extra spaces
Cleans up cells with lots of newlines and whitespace, preserving any newlines}
\usage{
extraspace(x)
}
\arguments{
\item{x}{The string to be converted}
}
\description{
Clean up extra spaces
Cleans up cells with lots of newlines and whitespace, preserving any newlines
}
