% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-functions.R
\name{firstclean}
\alias{firstclean}
\title{Generic cleaning functions
Some initial cleaning functions that are applied to all tables;
standardising hyphens, removing the markers for legislation updates}
\usage{
firstclean(table)
}
\arguments{
\item{table}{The table you have scraped from the page and now want to lean}
}
\description{
Generic cleaning functions
Some initial cleaning functions that are applied to all tables;
standardising hyphens, removing the markers for legislation updates
}
