% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flavourings}
\alias{flavourings}
\title{Flavourings}
\format{A dataset with 2504 rows and 9 columns}
\usage{
flavourings
}
\description{
A dataset containing the union list of approved flavourings
}
\keyword{datasets}
