% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-functions.R
\name{getbadtable}
\alias{getbadtable}
\title{Get table n from a html document (badly formatted)}
\usage{
getbadtable(htmldoc, n, ag = "ancestor")
}
\arguments{
\item{htmldoc}{the html document from the relevant page}

\item{n}{number of the table}

\item{ag}{table rows will be excluded that contain this string}
}
\description{
Get table n from a html document (badly formatted)
}
