% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-functions.R
\name{getnames}
\alias{getnames}
\title{Get names
Take a table, make the column names the top row, slice off the top row}
\usage{
getnames(table, n = 1)
}
\arguments{
\item{table}{The table you want to perform this function on}

\item{n}{The row number you want the names to be (default = 1)}
}
\description{
Get names
Take a table, make the column names the top row, slice off the top row
}
