% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-functions.R
\name{gettable}
\alias{gettable}
\title{Get table number n from webpage url}
\usage{
gettable(url, n)
}
\arguments{
\item{url}{URL of the page you want to get the table from}

\item{n}{number of the table}
}
\description{
Get table number n from webpage url
}
