% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geturls.R
\name{geturl}
\alias{geturl}
\title{Get latest EUR-Lex url for a piece of legislation}
\usage{
geturl(legurl)
}
\arguments{
\item{legurl}{The link to the legislation on legislation.gov.uk}
}
\description{
Get latest EUR-Lex url for a piece of legislation
}
