% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legurls}
\alias{legurls}
\title{Legislation URLs}
\format{A dataset with 8 rows and 3 columns}
\usage{
legurls
}
\description{
A dataset containing the location of the relevant legislation on
legislation.gov.uk and eur-lex
}
\keyword{datasets}
