% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-functions.R
\name{geturl}
\alias{geturl}
\title{Get the most up to date consolidated version from the legislation.gov.uk page}
\usage{
geturl(legurl)
}
\arguments{
\item{legurl}{URL of the legislation on legislation.gov.uk}
}
\description{
Get the most up to date consolidated version from the legislation.gov.uk page
}
